fix: 將文件範圍下推至向量查詢
合併 #591 的文件範圍向量查詢修正。

+++ src/main/java/org/km/llmwiki/search/embedding/EmbeddingProjectionReadinessRepository.java
@@ -429,7 +429,7 @@ public class EmbeddingProjectionReadinessRepository {
         Optional<String> latestFullStatus = latestFullStatus(workspaceId, corpus, target);
         boolean proof = generationApplied && targetOperationCompleted
                 && current.getProjectionSnapshotToken() != null && current.getFailedCount() == 0
-                && current.getIndexedCount() == current.getExpectedCount();
+                && current.getIndexedCount().intValue() == current.getExpectedCount().intValue();
         boolean missingTargetOperation = target > 0 && all.stream().noneMatch(r ->
                 r.getGeneration().longValue() == target);
         EmbeddingProjectionReadinessStatus next;
